addressing comments
- move cli parsing to a separate file
- change error message for unknown extra arguments
